style: openai provider config, *openai.Client 값도 받을 수 있게 처리
